back/requirements.txt
Django==3.2.*
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
django-cors-headers==3.11.0
django-filter==21.1
psycopg2-binary==2.8.6